Create nat-gateway-create-automatic.sh
This script will generate a NAT Gateway in your VPC and automatically associate ALL of your private subnets with it. This script will ONLY work if your private subnets are tagged "private" somewhere in the name (note line 10: Values=*private*). if you have a different naming convention, modify this line of code, or use the manual version of the script.