Create nat-gateway-destroy-manual.sh
This will destroy the NAT Gateway you just created using the manual script. This script is useful if you do not have "private" in the naming convention for your route tables, or if you want to be more granular in its application (as opposed to adding all private subnets, for instance).